requirements.txt
pytorch >= 1.1.0
torchvision
opencv-python
pillow
numpy
sigpyproc